refactor: move everything, standardize single file structure
