+++ aibtc-v1/components/agents_tab.py
@@ -1,5 +1,5 @@
-import streamlit as st
 import pandas as pd
+import streamlit as st
 
 
 def render_agents_tab():
