+++ aibtc-v1/components/tools_tab.py
@@ -7,10 +7,10 @@ from crewai_tools import Tool
 
 
 def get_tool_groups() -> Dict[str, Type]:
-    import aibtc_crews.tools
+    import crews.tools
 
-    importlib.reload(aibtc_crews.tools)
-    return aibtc_crews.tools.get_tool_groups()
+    importlib.reload(crews.tools)
+    return crews.tools.get_tool_groups()
 
 
 def extract_bun_run_command(source: str) -> str:
