--- app_v1/components/execution_tab.py
@@ -1,7 +1,0 @@
-import streamlit as st
-from components.crews.contract_audit_crew import render_contract_audit_crew
-
-
-def render_execution_tab():
-    st.write("This is where the magic happens.")
-    render_contract_audit_crew()
